off-chain/requirements.txt
# Core dependencies
flask==2.3.3
flask-cors==4.0.0
requests==2.31.0
websockets==10.4
aiohttp==3.8.5

# Machine Learning (simplified)
scikit-learn==1.3.0
pandas==2.0.3
numpy==1.24.3

# Database
redis==4.6.0

# Monitoring
prometheus-client==0.17.1

# Utilities
python-dotenv==1.0.0
loguru==0.7.0

